Sepalated AdditionalHeader class from curl.*
